# backend/docs/swagger.yaml
definitions:
...
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.TransactionHistoryAddressSummary:
    properties:
      city:
        type: string
      destination_area_id:
        type: string
      destination_area_label:
        type: string
      district:
        type: string
      full_address:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      phone_number:
        type: string
      postal_code:
        type: string
      province:
        type: string
      recipient_name:
        type: string
      title:
        type: string
      village:
        type: string
    type: object
  handler.TransactionHistoryDetail:
    properties:
      courier_name:
        type: string
      courier_service:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/handler.TransactionHistoryDetailItem'
        type: array
      order_id:
        type: string
      shipping_address:
        $ref: '#/definitions/handler.TransactionHistoryAddressSummary'
      shipping_address_id:
        type: integer
      shipping_cost:
        type: number
      shipping_status:
        type: string
      snap_token:
        type: string
      status:
        type: string
      total_amount:
        type: number
      tracking_number:
        type: string
      updated_at:
        type: string
    type: object
  handler.TransactionHistoryDetailItem:
    properties:
      id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      selected_color_hex:
        type: string
      selected_color_name:
        type: string
      selected_size:
        type: string
      subtotal:
        type: number
    type: object
  handler.TransactionHistoryDetailResponse:
    properties:
      data:
        $ref: '#/definitions/handler.TransactionHistoryDetail'
    type: object
  handler.TransactionHistoryListItem:
    properties:
      courier_name:
        type: string
      courier_service:
        type: string
      created_at:
        type: string
      id:
        type: integer
      order_id:
        type: string
      shipping_cost:
        type: number
      shipping_status:
        type: string
      status:
        type: string
      total_amount:
        type: number
    type: object
  handler.TransactionHistoryListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handler.TransactionHistoryListItem'
        type: array
      pagination:
        $ref: '#/definitions/handler.TransactionHistoryPagination'
    type: object
  handler.TransactionHistoryPagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.TransactionTrackingData:
    properties:
      biteship_order_id:
        type: string
      courier_name:
        type: string
      courier_service:
        type: string
      events:
        items:
          $ref: '#/definitions/service.ShippingTrackingEvent'
        type: array
      order_id:
        type: string
      raw_status:
        type: string
      shipping_status:
        type: string
      tracking_number:
        type: string
    type: object
  handler.TransactionTrackingResponse:
    properties:
      data:
        $ref: '#/definitions/handler.TransactionTrackingData'
    type: object
...
  service.ShippingTrackingEvent:
    properties:
      description:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Customer auth, product catalog, and admin CMS API for the Diagonals
    website backend.
  title: Diagonals API
  version: "1.0"
paths:
...
  /api/v1/transactions:
    get:
      description: Get paginated transaction history for the authenticated customer,
        optionally filtered by payment status
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by payment status (pending|paid|failed|refunded)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TransactionHistoryListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my transaction history
      tags:
      - Transactions
  /api/v1/transactions/{order_id}:
    get:
      description: Get one transaction detail by order id for the authenticated customer
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TransactionHistoryDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my transaction detail
      tags:
      - Transactions
  /api/v1/transactions/{order_id}/tracking:
    get:
      description: Get current shipping tracking data for a paid transaction. By default
        data is served from local transaction state; use refresh=true to fetch latest
        status and events from Biteship API.
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      - description: Refresh tracking from Biteship
        in: query
        name: refresh
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TransactionTrackingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my transaction tracking
      tags:
      - Transactions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"